add errors.py for exception definition
