+++ datajoint/connection.py
@@ -7,11 +7,12 @@ from contextlib import contextmanager
 import pymysql as client
 import logging
 from getpass import getpass
+from pymysql import err
 
 from . import config
-from . import DataJointError
+from .errors import DataJointError, server_error_codes
 from .dependencies import Dependencies
-from pymysql import err
+
 
 logger = logging.getLogger(__name__)
 
@@ -142,11 +143,12 @@ class Connection:
             else:
                 raise
         except err.ProgrammingError as e:
-            raise DataJointError("\n".join((
-                "Error in query:", query,
-                "Please check spelling, syntax, and existence of tables and attributes.",
-                "When restricting a relation by a condition in a string, enclose attributes in backquotes."
-            )))
+            if e.args[0] == server_error_codes['parse error']:
+                raise DataJointError("\n".join((
+                    "Error in query:", query,
+                    "Please check spelling, syntax, and existence of tables and attributes.",
+                    "When restricting a relation by a condition in a string, enclose attributes in backquotes."
+                ))) from None
         return cur
 
     def get_user(self):
