+++ datajoint/external.py
@@ -1,6 +1,7 @@
 import os
 from tqdm import tqdm
-from . import config, DataJointError
+from . import config
+from .errors import DataJointError
 from .hash import long_hash
 from .blob import pack, unpack
 from .base_relation import BaseRelation
