+++ datajoint/jobs.py
@@ -3,7 +3,7 @@ from .hash import key_hash
 import os
 import pymysql
 from .base_relation import BaseRelation
-from . import DataJointError
+from .errors import DuplicateError
 
 ERROR_MESSAGE_LENGTH = 2047
 TRUNCATION_APPENDIX = '...truncated'
@@ -85,7 +85,7 @@ class JobTable(BaseRelation):
             user=self._user)
         try:
             self.insert1(job, ignore_extra_fields=True)
-        except (pymysql.err.IntegrityError, DataJointError):
+        except DuplicateError:
             return False
         return True
